movietic/fastapi_service/requirements.txt
fastapi==0.128.0
uvicorn==0.30.0
razorpay==1.4.2
pymongo==4.7.0
python-dotenv==1.0.1
PyJWT==2.10.1
